feat: setup env

+++ product-service/models/product_model.go
@@ -4,7 +4,7 @@ import "gorm.io/gorm"
 
 type Product struct {
 	gorm.Model
-	Name        string `json:"name"`
-	Price       int    `json:"price"`
-	Description string `json:"quantity"`
+	Name        string  `json:"name"`
+	Price       float64 `json:"price"`
+	Description string  `json:"quantity"`
 }
